backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg2-binary>=2.9.7
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==1.10.9
email-validator==1.3.1
bcrypt==3.2.2